added a hack in HumanoidDynamicRobot to address the following issue: The HRP2Specificities.xml file is used to identify  groups of joints, like arm joints, leg joints, etc. In the current version of this file, the issue lies in the arms groups. Here is the identification of the right arm joints:
<Arms>
    <Right>
        <UpperArmLength> 0.25 </UpperArmLength>
        <ForeArmLength> 0.25 </ForeArmLength>
        <JointNb> 7 </JointNb>
        <JointsID> 16 17 18 19 20 21 22 </JointsID>
    </Right>

The problem is that the seventh joint (joint 22) indicates the the opening of the hand and should be separated from the kinematic chain of the arm.

Changing the the value of the joint numbers from 7 to 6, and/or removing the last joint number(22) made a mess in the building of the robot. So for now it is kept as it is. Instead of that, this update fixes the problem with a hack in HumanoidDynamicMultibody.cpp:

void HumanoidDynamicMultiBody::LinkBetweenJointsAndEndEffectorSemantic()

original code commented out.