1/ SpiritVRMLReader.cpp: Adding rules to improve the grammar of the parser.
Conclusion: this is probably the final modification
on this parser. Next generation will be separated
from dynamicsJRLJapan, will rely on openvrml which
implements a full VRML 97 grammar, and will use
the abstract interface.

2/ TestHumanoidDynamicRobot.cpp: beautification of
the output for testing. Still not complete.

3/ sample.wrl: Modification of the model to avoid
parsing the geometrical part (currently the part
where the current grammar fails).