Fix compilation pb + compensate for initial rotation.
This commit fix a compilation pb for model generation
targeted to AMELIF.
It also fix a problem for the initial rotation of
objects.
Now
./CreateVRMLPoseFile $ROBOTPATH robotmodel.wrl \
robotspecificites.xml robotlinkrank.xml position.articular
generates a VRML-2 compliant model for windows-Blender.

On HRP2JRL, there is still a problem with the left hand.
The current program assume one VRML file per body.
A small reorganization of the VRML files is necessary for
a correct import with Windows-Blender.