Fix Euler angles computation and validation against HRP-2 10/14.
With this commit it is now possible to generate automatically
the inertia matrix using Maple. The code generated can be efficiently
compiled and is usually 30 % faster than the generic algorithm.

To generate the maple description:
./CreateMapleModel PATH_TO_MODEL MODEL_NAME SPECIFICITIES_FILE
LINK_JOINT_RANK position.articular

This generate 4 files:
ROBOTKinematicData.maple
ROBOTDynamicData.maple
ROBOTAdditionalData.maple
ROBOT.shmaple

To generate the C-code assuming that you have maple and
ModelGeneration.maple from Humans:
./ROBOT.shmaple
(once it is executable)

We might have efficiency decrease due to the fact that currently
Pi is not handle as a symbol but as a numerical value.

ModelGeneration.maple has been written by P.-B. Wieber.