Fix problem for parsing *Bush* models.
This commit rationalize the way bodies and joints
are build according to the models.
*Bush* models (at least HRP-2 one's) have complex
joints structure without bodies.
A body is however introduced in the robot model
to deal with those complex structures. The body
is massless, and is internally named a virtual
body.

Slide joints are ignored in this commit, but will
be deal with in future commits.