Roll back previous modifications on gravity.
After more reading on Featherstone Rigid Body Dynamics and HoR,
it appears that for few days I totally forgot that we are doing
control...
The RENA computes the force and the torques needed to realize
a given [q,dq,ddq]. It is thus normal to have a positive
force on the free flyer. It means that the robot is at the
equilibrium at that we compensate for the external forces
(i.e. gravity when the robot is not moving).