Fix pbs related to previous commit.
Please do not push to the master a branch which do not pass
the tests.
In this specific case, the abstract interface did not work
because the new objects fields were not properly initialized.
The JointPrivate class needs to resize the related body's
spatial representation whenever the number of Dofs is changed.
Therefore a new methode setnumberDof() was introduced and
m_nbDofs is now private.

Another problem is the static cast of the related body.
A shared pointer was introduced to allow a carefull
management of the objects, but the class DynamicBody encapsulating
this shared pointer is the one returned by linkedBody().
Therefore there is an indirect access to DynamicBodyPrivate whenever the tree
is build from scratch as it is in AbstractInterface.
Some static cast remains, they will be removed later on.

Please be aware that crashing the master breaks the buildbot
nightly generation of binary packages. It also send a bad signal
to whoever is trying our code.